feat(userTest): componentized the audio recording for user test.
